Revert "2,4ghz channel 6 => 13"
This reverts commit 24145e18b372e1f42f7e1f7a50706edd274fa090.

The option isn't (yet) automatically changed on upgrade, thus this
wouldn't propagate to already installed gluon routers. Also, channel 13
may not be supported by all devices.

We can do this change when we can specify own package feeds, write a
package that sets the channel on every upgrade and know that the
majority of routers support the channel.