Add gluon repository as submodule
Instead of recording the commit in the gluon repository that belongs to
a release of our firmware in the README.md, it is more reasonable to
have the gluon repository as a submodule, having this information in a
machine-readable way and making it easier to reproduce the compilation
of a specific version.

To compile gluon, simply add `GLUON_SITEDIR="$(dirname $PWD)"` to the
make command.