use port 50000 for batman v15 on vpn04
This is supposed to be the only server to support VPN connections for
batman v15 on port 50000, as fallback for nodes that cannot use the new
VPN port 50001 due to restrictive firewall settings.

Since this server has not been used in any stable or testing firmware so
far, there is no risk that batman v13 nodes will try to connect to this
port.